requirements.txt
appdirs==1.4.4
argcomplete==1.12.0
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorlog==4.1.0
distlib==0.3.1
filelock==3.0.12
idna==2.10
isort==5.0.9
mypy==0.782
mypy-extensions==0.4.3
nox==2020.5.24
pathspec==0.8.0
py==1.9.0
pygal==2.4.0
python-dotenv==0.14.0
regex==2020.6.8
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
virtualenv==20.0.26
